Introducing Planck and Stoney natural units. (#26)
These units are in such extreme ranges that it required an overhaul to how I've
been doing approximation.  Before, I was being kind of scattershot with
absolute differences and having to specify many more `within=` tolerances.  Now
it's based on the _ratio_ between the two quantities after conversion, which
can scale better.  I've only needed to supply tolerances for a small handful of
unit tests, which are probably due to approximations on Wikipedia.

The Planck and Stoney tests confirm that we can preserve some fairly high
precision with the floating point conversions, and memorialize that fact in
their test suites.

This also required an additional stage of unit conversion, where we attempt to
"cancel" out any factors in the remaining units.  For example, if after
collapsing dimensions, we're left with Second * Ampere, we know we can reduce
that to a simpler dimension Coulomb by actually performing the multiplcation.