Correcting an issue where prefixes weren't round-tripping correctly.
This was a persistent issue with the test suite, randomly failing a hypothesis
example occasionally.  I had some confusion in the string formatting for
prefixed units, where the precedence would get out of whack.  For example, when
formatting `(Kilo*Meter)**2`, we'd format that as `Mm²` because a square
kilometer is a million square meters; however the parsing would treat that as
if it were a `(Mega*Meter)**2`.  There were also a few more issues and a test
flake in MathML formatting.